Make classical incremental GC version compilable again

+++ rts/motoko-rts/src/memory.rs
@@ -15,10 +15,10 @@ use crate::constants::GB;
 // TODO: Redesign for 64-bit support by using a dynamic partition list.
 /// Currently limited to 64 GB.
 #[enhanced_orthogonal_persistence]
-pub const MAXIMUM_MEMORY_SIZE: Bytes<usize> = Bytes(64 * GB);
+pub const MAXIMUM_MEMORY_SIZE: Bytes<u64> = Bytes(64u64 * GB as u64);
 
 #[classical_persistence]
-pub const MAXIMUM_MEMORY_SIZE: Bytes<usize> = Bytes(usize::MAX);
+pub const MAXIMUM_MEMORY_SIZE: Bytes<u64> = Bytes(usize::MAX as u64);
 
 
 // Memory reserve in bytes ensured during update and initialization calls.
